Refactor fixes to be pure & auto derive checks from fix outputs
This completely changes how checking & fixing is
implemented. Previously, we ran fixes outside the sandbox in the fix
script, and we skipped fixing if a corresponding check passed...

Now, we build a derivation that computes the patch for the fix, which
we use in the fix script to apply the fix. With this new approach, fix
outputs are now cached and are computed in parallel!

Also, now we automatically derive check commands from fix outputs. If
there are any diffs (or if the fix command returned a non-zero exit
code), the check will fail, otherwise it'll pass.